Added Loader Script and DebugManager Moved Debug tools from GameManager to DebugManager Removed GameManager from the scene
