+++ SimpleFighter/Assets/Scripts/DebugManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 653adc39468a1fb45b948b2923b779a5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
