+++ SimpleFighter/Assets/Scripts/GameManager.cs
@@ -6,42 +6,36 @@ using UnityEngine.UI;
 public class GameManager : MonoBehaviour
 {
 
-    public Text fpsText;
-    public float deltaTime;
+    public static GameManager instance = null;
+    
 
     // Start is called before the first frame update
-    void Start()
+    void Awake()
     {
+        //Check if instance already exists
+        if (instance == null)
+        {
+            //if not, set instance to this
+            instance = this;
+        }
+        //If instance already exists and it's not this:
+        else if (instance != this)
+        {
+            //Then destroy this. This enforces our singleton pattern, meaning there can only ever be one instance of a GameManager.
+            Destroy(gameObject);
+        }
+
+        //Sets this to not be destroyed when reloading scene
+        DontDestroyOnLoad(gameObject);
+        
         //cap the frame rate at 60fps
         Application.targetFrameRate = 60;
-        
     }
 
     // Update is called once per frame
     void Update()
     {
-
-
-		//DEBUG TOOLS
-
-        //fps display here
-        deltaTime += (Time.deltaTime - deltaTime) * 0.1f;
-        float fps = 1.0f / deltaTime;
-        fpsText.text = "fps " + Mathf.Floor(fps).ToString();
-
-        //PAUSE & UNPAUSE
-        if (Input.GetKeyDown(KeyCode.BackQuote))
-        {
-            if (Time.timeScale == 1.0f)
-                Time.timeScale = 0.0f;
-            else
-                Time.timeScale = 1.0f;
-            // Adjust fixed delta time according to timescale
-            // The fixed delta time will now be 0.02 frames per real-time second
-            Time.fixedDeltaTime = 0.02f * Time.timeScale;
-        }
-
-
-
+		
+     
     }
 }
