+++ SimpleFighter/Assets/Scripts/Loader.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 771b1cb4ee8d6f8478968cd7e56c58cd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
